chore(designer-frontend): move usePreviewContext from app-development to shared
usePreviewContext is consumed by packages/ux-editor and
packages/ux-editor-v4, which had to import it directly from the
app-development workspace package. Move PreviewContext into
packages/shared so packages depend on the shared layer instead of
reaching into an app-level package.

Closes #14304
