+++ src/Designer/frontend/app-development/layout/PageHeader/SubHeader/PreviewButton/PreviewButton.test.tsx
@@ -6,7 +6,7 @@ import { PageHeaderContext } from 'app-development/contexts/PageHeaderContext';
 import { renderWithProviders } from 'app-development/test/mocks';
 import { PackagesRouter } from 'app-shared/navigation/PackagesRouter';
 import { pageHeaderContextMock, previewContextMock } from 'app-development/test/headerMocks';
-import { PreviewContext } from 'app-development/contexts/PreviewContext';
+import { PreviewContext } from 'app-shared/contexts/PreviewContext';
 import { app, org } from '@studio/testing/testids';
 
 jest.mock('@studio/components-legacy/src/hooks/useMediaQuery');
