+++ src/Designer/frontend/app-development/test/headerMocks.ts
@@ -4,7 +4,7 @@ import { HeaderMenuItemKey } from 'app-development/enums/HeaderMenuItemKey';
 import { type HeaderMenuItem } from 'app-development/types/HeaderMenu/HeaderMenuItem';
 import { userMock } from './userMock';
 import { type StudioProfileMenuGroup, type StudioProfileMenuItem } from '@studio/components';
-import { type PreviewContextProps } from 'app-development/contexts/PreviewContext';
+import { type PreviewContextProps } from 'app-shared/contexts/PreviewContext';
 import { org, app } from '@studio/testing/testids';
 
 const menuItemsMock: HeaderMenuItem[] = [
