+++ src/Designer/frontend/packages/ux-editor-v4/src/containers/DesignView/AddItem/hooks/useBaseAddComponentHandler/useBaseAddComponentHandler.test.tsx
@@ -5,10 +5,10 @@ import { ComponentType } from 'app-shared/types/ComponentType';
 import { BASE_CONTAINER_ID } from 'app-shared/constants';
 import { useBaseAddComponentHandler } from './useBaseAddComponentHandler';
 import { waitFor } from '@testing-library/react';
-import { usePreviewContext } from 'app-development/contexts/PreviewContext';
+import { usePreviewContext } from 'app-shared/contexts/PreviewContext';
 import { useFormItemContext } from '../../../../FormItemContext';
 
-jest.mock('app-development/contexts/PreviewContext');
+jest.mock('app-shared/contexts/PreviewContext');
 jest.mock('../../../../../hooks', () => ({
   useAppContext: () => ({
     selectedFormLayoutSetName: 'testLayoutSet',
